fix(ec2): fix cross-stack SecurityGroup connections (#1678)
Automatic cross-stack SecurityGroup connections did not work because
they would create objects in both stacks, both of which would refer
to the other stack. This would lead to a cyclic dependency.

After this change, if the SGs are detected to be in different stacks,
both Ingress/Egress Rule objects will be situated in the stack of
the object on which the 'allowFrom'/'allowTo' call is made.

The behavior is only different for cross-stack relationships to
keep disruption to deployed stacks to a minimum.

Fixes #1654.