feat(aws-eks): add construct library for EKS (#1655)
Construct library to set up an EKS cluster and add nodes to it.

Generalizes adding AutoScalingGroup capacity and make it the same
between both ECS and EKS clusters. Fixes naming inconsistencies
among properties of `AutoScalingGroup`, and between EC2 AutoScaling and
Application AutoScaling.

This PR takes #991 but reimplements the API in the style of the ECS library.

BREAKING CHANGE: For `AutoScalingGroup`, renamed `minSize` =>
`minCapacity`, `maxSize` => `maxCapacity`, for consistency with
`desiredCapacity` and also Application AutoScaling.

For ECS's `addDefaultAutoScalingGroupCapacity()`, `instanceCount` =>
`desiredCapacity` and the function now takes an ID (pass
`"DefaultAutoScalingGroup"` to avoid interruption to your deployments).