fix(ssm): Generate correct SSM Parameter ARN
The SSM parameter name (`resourceName` portion of it's ARN) is required
to start with a `/`, which caused the presence of a doubled `//` in the generated
ARN caused by the fact `sep` defaults to `/`.

This change allows `sep` to be set to an empty string, allowing SSM rendering
of correct SSM parameter ARNs.