feat(lambda-event-sources): "api" event source (#1742)
Introduce an "api" lambda event source which will implicit create an
API Gateway API and connect the resource in the specified path + method
to the Lambda target.

Added `resourceForPath` which creates a new resource (and all intermediate 
resources) mounted at a specified path in the API model.

Added `getResource` can be used to find a resource by path part.

Introduce an abstract `ResourceBase` which is used to share the implementation
between the root resource and all other resources.

Convert HTTP methods to uppercase.