fix: Npm ignores files and folders named "core" by default
Since the release of `npm@6.8.0`, a version of `npm-packlist` is
included that introduces a behavior to ignore anything named "core" by
default.

Added `!core` to the `.npmignore` file as a work around, and also more
strictly pinned the `npm` version to auto-upgrade to so we don't get
surprised in this kind of way anymore.

Cause: npm/npm-packlist#24