feat(toolkit): store context in cdk.context.json
Separate out context into a separate file (`cdk.context.json`) to make
it easier for users to decide how to treat as part of their version
control story.

*We still recommend that you check in `cdk.context.json`*, but you may
decide for whatever reason that this does not work for you, and in those
cases you can add the new file to your `.gitignore` pattern.

For backwards compatibility, if context was loaded from `cdk.json`, it
will be written  back to `cdk.json`, but if not it will be stored in the
new file instead.

Fixes #1662.