fix(aws-codepipeline): update CFN example. (#1653)
This test is referred as example in the
[Cloudformation README](https://github.com/awslabs/aws-cdk/blob/master/packages/%40aws-cdk/aws-cloudformation/README.md).
But the example doesn't do what it intended.
Without specifying the runOrder the `pipelineExecuteChangeSetAction` will be triggered immediately without change set being populated,
which will lead to deployment failure.
After adding the `runOrder` the 3 actions will run in order and only be deployed after manual approval.