Upgrade to CloudFormation Resource Specification v2.24.0
Added new namespace packages:
- `@aws-cdk/aws-apigatewayv2` for `AWS::ApiGatewayV2`
- `@aws-cdk/aws-fsx` for `AWS::FSx`
- `@aws-cdk/aws-ram` for `AWS::RAM`
- `@aws-cdk/aws-robomaker` for `AWS::Robomaker`

Had to tweak the code generation to support a new feature of the
`AWS::CodeBuild::Project` resource that uses an unusual model (list of
list of some property type), and the new code should support a broader
range of `PropertyType` definitions. This also supports using
`PropertyTypes` to alias other types (for example, `String`), as has
been seen to be done in the `AWS::EC2` namespace for example.

Removed test cases (enforcing the presence of a documentation link,
having a `Required` flag, having an `UpdateType`) as the specification
appears to assume defaults for those.