feat(core): democratize synthesis
This change enables each construct in the tree to participate in synthesis
by overriding the "synthesize" method and emitting files to the session directory.

It is a small step towards generalizing how synthesis artifacts are being
produced, built and deployed by the CDK.

This change moves the code that synthesizes a CloudFormation template from a 
CDK stack to the `Stack` class and emits a `.template.json` file for each
stack.

The CLI hasn't been changed to work with these new files yet, so the output
of this file is ALSO included in the main manifest (defined `cxapi.OUTFILE_NAME`),
which is the only file currently being read by the CLI.