chore: Upgrade to CloudFormation Resource Specification v2.24.0 (#1886)
Added new namespace packages:
- `@aws-cdk/aws-fsx` for `AWS::FSx`
- `@aws-cdk/aws-ram` for `AWS::RAM`
- `@aws-cdk/aws-robomaker` for `AWS::RoboMaker`

Had to tweak the code generation to support a new feature of the
`AWS::CodeBuild::Project` resource that uses an unusual model (list of
list of some property type), and the new code should support a broader
range of `PropertyType` definitions. This also supports using
`PropertyTypes` to alias other types (for example, `String`), as has
been seen to be done in the `AWS::EC2` namespace for example.

Removed test cases (enforcing the presence of a documentation link,
having a `Required` flag, having an `UpdateType`) as the specification
appears to assume defaults for those.

And since this tried to introduce two "V2" packages, took the
opportunity to make changes such that the "V2" resources are
colocated with their "V1" counterparts, albeit with a "V2" appended
to the generated class names.

Fixes #1888