feat(toolkit): store context in cdk.context.json  (#1774)
Separate out context into a separate file (`cdk.context.json`) to make
it easier for users to decide how to treat as part of their version
control story.

*We still recommend that you check in `cdk.context.json`*, but you may
decide for whatever reason that this does not work for you, and in those
cases you can add the new file to your `.gitignore` pattern.

Context currently in `cdk.json` will be migrated to `cdk.context.json` 
as it is updated. For backwards compatibility, the `cdk context` 
command will continue to manage both types of context, but at some
point in the future future will only manage the context in the separate file.

Fixes #1662.