feat(route53): Convenience API for creating zone delegations (#1853)
When creating delegation relationship between two `PublicHostedZone`s,
one can now use `zone.addDelegation(otherZone)` instead of manually
creating the `ZoneDelegationRecord` insteance. This reduces the risk of
passing the incorrect name server, or hosting the record on the wrong
end of the relationship (DNS is hard!)

Additionally, fixes a bug in which it was not possible to create a zone
delegation record using a `IHostedZone.hostedZoneNameServers` property
as the array was mapped, which caused the `Fn::GetAtt` stringified list
token to become corrupted.

Fixes #1847