feat(core): can use Constructs to model collections of Stacks
Stacks no longer accept a `cdk.App` as scope, but a `cdk.Construct` like
other constructs. This makes it possible define a Construct to
represent your entire application, which can consist of a collection of
Stacks.

Stack names will automatically be derived from their location in the
construct tree, but it's also possible to specify the deployed name of
the stack by specifying the `stackName` property.

Calling `app.run()` is no longer necessary, but can still safely
be done.

Fixes #1479.