build: Allow using jest as a test framework (#1955)
If a package's `package.json` file contains a `jest` entry (used for
configuring the behavior of `jest`), assert the package uses `jest`
instead of `nodeunit` as it's test harness (and fail if there are any
`nodeunit` tests - based on file name).

This is going to make it easier to migrate away from `nodeunit`.