Make instanceId optional
`instanceId` is an arbitrary identifier for an instance registered to a
Cloud Map service. It is not particularly useful to the customer, but
due to science it is required on a RegisterInstance API call. This makes
instanceId an optional prop as it was meant to be on the instance
constructs.