chore: tag AWS packages with docs information
This PR starts the enhancing of the AWS packages with title and tag
information. The title will be used to render docs pages, and the tags
will be used to group packages by workload/type so that we can generate
helpfully grouped overview pages out of our 100+ services.

Please note that at this moment I care more about a place for this
information to live than the 100% accuracy of these tags. I bet many
of them can and will be debated, and once merged you're free to
submit changes to any of them. I do not want to start a discussion
about the exact tags applied, but about the metadata itself.