feat(cloudformation): allow specifying additional inputs for deploy Actions (#2020)
Additional input Artifacts are needed when using the `parameterOverrides` property of the Action (any Artifact used in that map needs to be added to the input Artifacts of the Action, otherwise the Pipeline will fail at runtime).

Also did some cleanup in the Action superclass while I was in the area.

Fixes #1247