fix(cfn2ts): properly de-Tokenize L1 string-arrays (#2033)
Because of accepted Tokenized values, cfn2ts previously translated a
string and string array into:

    interface Props {
        someString: string;
        stringArray: Array<string | Token> | Token;
    }

The latter was not intended. We now emit:

    interface Props {
        someString: string;
        stringArray: string[];
    }

Since the Token will be typed as `Object` in Java, this change now makes
it impossible to mistakenly pass any old java Object. A typical case
would be pass a `Subnet` where a `string subnetId` was expected, failing
in a nonobvious way.

Fixes #2030.