fix(aws-ecs): handle long ARN formats for services (#2176)
Due to a bug in Cloudformation, Fn::GetAtt for service name on an ECS
service was returning the cluster name instead of the service name. This
was causing the ResourceId on
AWS::ApplicationAutoScaling::ScalableTarget to be set incorrectly.

This change forces the customer to specify whether they have opted into
the long ARN format for services (which can be achieved through the following command:
`aws ecs put-account-setting-default --name serviceLongArnFormat --value
enabled`). The CDK cannot currently detect if the customer has the these
settings enabled so there could be a potential conflict on deployment.

Fixes #1849.