fix(lambda): avoid OperationAbortedException when using log retention
If multiple LogRetention constructs are present in the stack, they will try to
act on the provider's log group at the same time. This can sometime result in
an OperationAbortedException. To avoid this and because this operation is not
critical it is better to catch all errors when acting on the provider's log
group