Not finished....
This branch contains initial implementation of serialization/deserialization scheme
for the CDK, but we haven't finished it because it appears that the use case
for actually using CFN imports/exports within the same account/region
is not something too many customers need, especially given the implicit
references we have inside a CDK app.

There are use cases for being able to transmit complex constructs
between accounts/regions/environments, but if we think about these
use cases more broadly, the current implementation is lacking and there
are more things to consider such as protocl versions, etc.

All in all, this is junk. It's quite nice, but still, not useful enough
to persue at this point.