feat: use IVirtualService when adding backends
When adding backends to a node, now we use IVirtualService instead fo
virtualServiceName, this make sit easier to catch errors and reduce
testing. With thiis it's verified that the correct type is passed in
and not just a string containing any ordinary value. The unit and
integration tests were also updated.