feat(config): add constructs for managed and custom rules
Specific classes are exposed for an AWS managed rule (`ManagedRule`) and a custom rule
(`CustomRule`) with support for compliance and re-evaluation events.

Higher level constructs for configured AWS managed rules are also exposed (starting with Access
Keys Rotated and CloudFormation rules). This defines a pattern for future configured managed rules.