fix: implement IResource and Resource
This will probably come back, there seem to be chnges going on in the
core repo. Though the only way I found to ignore the UID property was to
add awslint rule to ignroe it.
The UID property seems useless for now and makes the code harder too read.
I also added awslit rules to ignore the ImportedXZYProps as it caused
other issues with the export method.

This will most likely be rejected, I will stop making changes until instructed
to otherwise, I just wanted to get a working build.