refactor(awslint): XxxReflection classes
Implement a "Reflection" class for each rule family (modules, constructs, resources) which
reflect on the type system and expose a relevant API for linter rules.

The rational is that this increases our ability to reuse the model by other rule families.
For example, the "ResourceReflection" extends "ConstructReflection" and adds more
information that's specific to AWS resource constructs.