feat(elbv2): add fixed response support for application load balancers (#2328)
It follows the same pattern as the one used for `addTargetGroups` with the check on `priority`.

This allows for the following scenario: an infrastructure stack exporting an application listener with
a default fixed response and multiple application stacks registering their services with specific
paths/headers on the imported application listener.