feat(cdk): support encoding Tokens as numbers (#2534)
Numbers can now be encoded into a set of (hugely negative) numbers,
and the encoding can be reversed.

This allows APIs that take numbers to take lazy values and intrinsics
that evaluate to numbers.

This change only introduces the capability, it does not use it in any
of the construct libraries yet.

Fixes #1455.