feat(cli): cloud assembly - decouple synth and deploy
formalize the concept of a cloud assembly to allow decoupling "synth" and "deploy". the main
motivation is to allow "deploy" to run in a controlled environment without needing to execute the app for security purposes.

"cdk synth" now produces a self-contained assembly in the output directory, which we call a "cloud assembly". this directory includes synthesized templates (similar to current behavior) but also a "manifest.json" file and the asset staging directories.

"cdk deploy -a <assembly-dir>" will now skip
synthesis and will directly deploy the assembly.

to that end, we modified the behavior of asset staging such that all synth output always goes to the output directory, which is by default `cdk.out` (can be set with `--output`). if there's a single template, it is printed to stdout, otherwise the name of output directory is printed.

BREAKING CHANGE: apps created with this version cannot be deployed using a previous version of the CLI and vice versa. Both have to be >= 0.32.0.

- @aws-cdk/cdk.App member synthesizeStack has been removed
- @aws-cdk/cdk.App member synthesizeStacks has been removed
- @aws-cdk/cdk.FileSystemStore has been removed
- @aws-cdk/cdk.InMemoryStore has been removed
- @aws-cdk/cdk.SynthesisSession member addBuildStep has been removed
- @aws-cdk/cdk.SynthesisSession member store has been removed
- @aws-cdk/cdk.FileSystemStoreOptions has been removed
- @aws-cdk/cdk.ISessionStore has been removed
- @aws-cdk/cdk.ISynthesisSession member addBuildStep has been removed
- @aws-cdk/cdk.ISynthesisSession member store has been removed
- @aws-cdk/cdk.ManifestOptions member legacyManifest has been removed
- @aws-cdk/cdk.ResolveContext has been removed
- @aws-cdk/cdk.SynthesisOptions member legacyManifest has been removed
- @aws-cdk/cdk.SynthesisOptions member store has been removed
- @aws-cdk/cx-api.BuildManifest has been removed
- @aws-cdk/cx-api.BuildStep has been removed
- @aws-cdk/cx-api.SynthesizeResponse has been removed
- @aws-cdk/cx-api.SynthesizedStack has been removed
- @aws-cdk/cx-api.BuildStepType has been removed
- --interactive has been removed
- --numbered has been removed