feat(tokens): enable type coercion
relax restrictions on input types for Token.toXxx in order to allow
flexible type coercion.

this may be needed in situations where users want to force a token
typed as one type to be represented as another type and generally
allow tokens to be used as "type-system escape hatches".

Previously, this did not work:

    const port = new Token({ "Fn::GetAtt": [ "ResourceId", "Port" ] }).toString(); 
    new TcpPort(new Token(port).toNumber());

Also, this did not work:

    const port = new Token({ "Fn::GetAtt": [ "ResourceId", "Port" ]}).toNumber();

Fixes #2679