chore(build): Disable concurrency in pack.sh
There appear to be a file system race condition during dotnet project
builds where parallel instances race for the same files, resulting in
sporadic build failures due to not being able to access a file (that was
cleaned up by another process already).

Disabling parallelism seems to be the easiest way forward, and also the
cheapest way to validate the root cause was correctly identified.