feat: physical names in the entire Construct Library.
This PR changes the xyzName attribute of all Resources in the Construct Library to be of type PhysicalName,
and adds an AWS linter rule that checks ensures conformance.

The name of the property can be any ending substring of the base name of the class with the Name suffix -
for example, if my resource is AwesomeFoobar, the name can be foobarName or awesomeFoobarName
(that's because we often have Specialized1AwesomeFoobar and Specialized2AwesomeFoobar in our library,
that share a base prop interface).

BREAKING CHANGE: all <resource>Name attributes have their type changed from string to cdk.PhysicalName.