feat(lambda): Expose $LATEST function version (#2792)
Sometimes, one wants to create an `Alias` to the `$LATEST` version of a
`Function`, but this requires access to an `IVersion` representing that.
This features adds a `latestVersion` property to `IFunction` that offers
simple access to `$LATEST`. Prior to this, each user needing to access
this would have had to roll theiw own implementation of `IVersion`,
which is rather cumbersome.

Fixes #2776